event_api_server/requirements.txt
fastapi==0.88.0
kafka_python==2.0.2
uvicorn==0.20.0
psycopg==3.1.4
psycopg_pool==3.1.4
pydantic==1.10.2

